added delete functionality
added delete buttons, but they don't work properly : it's possible to delete element only by clicking the first button: elements are deleted starting from the first one. they are deleted from client as button is pressed, but deleted from the bottom to top of the list on the server.
to-do : fix bug and polish design